requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.7.0
billiard==3.5.0.4
celery==4.2.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3
cssselect==1.0.3
Django==2.1
django-rest-swagger==2.2.0
djangorestframework==3.8.2
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10
kombu==4.2.1
lxml==4.2.4
MarkupSafe==1.0
openapi-codec==1.3.2
parsel==1.5.0
pkg-resources==0.0.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
pytz==2018.5
queuelib==1.5.0
redis==2.10.6
requests==2.19.1
Scrapy==1.5.1
scrapy-redis==0.6.8
service-identity==17.0.0
simplejson==3.16.0
six==1.11.0
Twisted==18.7.0
uritemplate==3.0.0
urllib3==1.23
vine==1.1.4
w3lib==1.19.0
zope.interface==4.5.0
